# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aegis AI — Backend REST API",
    "version": "1.0.0",
    "description": "Generated from the running Express router by `scripts/openapi.mjs`; the\nroute inventory is never hand-maintained. Prose lives in\n`scripts/openapi.meta.mjs`.\n\nRoutes marked `x-documented: false` are served but not yet described.\n\nEvery route below is also reachable under `/api/*` without the version\nsegment. That alias is the legacy spelling of this same router, kept for\nbackward compatibility — not a second API.\n\nThis document is generated to a file and served nowhere. A complete route\ninventory maps the attack surface."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jwt"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "SuccessEnvelope": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "success"
            ]
          },
          "data": {
            "description": "The payload; shape varies by endpoint."
          },
          "results": {
            "type": "integer",
            "description": "Row count, on list responses only."
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  },
  "paths": {
...
    "/api/v1/enterprise/renewals": {
      "get": {
        "tags": [
          "Enterprise Console"
        ],
        "summary": "Renewals raised, and renewals nobody has noticed yet.",
        "x-permission": "policy.read",
        "x-realm": "ENTERPRISE",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed; the offending parameter is named."
          },
          "401": {
            "description": "No valid session."
          },
          "403": {
            "description": "Authenticated, but not permitted."
          }
        }
      }
    },
...
    "/api/v1/intelligence/analytics/renewals": {
      "get": {
        "tags": [
          "Intelligence"
        ],
        "summary": "Renewal exposure across the book.",
        "x-permission": "analytics.read",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed; the offending parameter is named."
          },
          "401": {
            "description": "No valid session."
          },
          "403": {
            "description": "Authenticated, but not permitted."
          }
        }
      }
    },
...
    "/api/v1/knowledge/route": {
      "post": {
        "tags": [
          "Knowledge"
        ],
        "summary": "Find the guidance that answers a question.",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed; the offending parameter is named."
          },
          "401": {
            "description": "No valid session."
          },
          "403": {
            "description": "Authenticated, but not permitted."
          }
        }
      }
    },
...
    "/api/v1/leads": {
      "get": {
        "tags": [
          "Leads"
        ],
        "summary": "The pipeline, paginated.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "1-based page number.",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Rows per page; capped at 100.",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "x-permission": "lead.read",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed; the offending parameter is named."
          },
          "401": {
            "description": "No valid session."
          },
          "403": {
            "description": "Authenticated, but not permitted."
          }
        }
      },
      "post": {
        "tags": [
          "Leads"
        ],
        "summary": "Raise a lead.",
        "description": "Any signed-in account may raise one; reading the pipeline is a capability.",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed; the offending parameter is named."
          },
          "401": {
            "description": "No valid session."
          },
          "403": {
            "description": "Authenticated, but not permitted."
          }
        }
      }
    },
    "/api/v1/leads/{id}": {
      "delete": {
        "tags": [
          "Leads"
        ],
        "summary": "Delete a lead.",
        "description": "Requires fresh authentication.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "x-permission": "lead.delete",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed; the offending parameter is named."
          },
          "401": {
            "description": "No valid session."
          },
          "403": {
            "description": "Authenticated, but not permitted."
          }
        }
      },
      "get": {
        "tags": [
          "Leads"
        ],
        "summary": "One lead.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "x-permission": "lead.read",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed; the offending parameter is named."
          },
          "401": {
            "description": "No valid session."
          },
          "403": {
            "description": "Authenticated, but not permitted."
          }
        }
      },
      "put": {
        "tags": [
          "Leads"
        ],
        "summary": "Update a lead.",
        "description": "Requires fresh authentication.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "x-permission": "lead.write",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed; the offending parameter is named."
          },
          "401": {
            "description": "No valid session."
          },
          "403": {
            "description": "Authenticated, but not permitted."
          }
        }
      }
    },
...
    "/api/v1/renewal-leads": {
      "get": {
        "tags": [
          "Renewal queue"
        ],
        "summary": "The renewal request queue, most urgent first.",
        "description": "Needs `lead.read`. `?format=csv` exports the page through the platform's report writer, which quotes every field and neutralises a leading formula character. Filters: `status`, `channel`. Not tenant-scoped — a consumer's organisationId is null by design, so these rows belong to the Aegis consumer funnel.",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed; the offending parameter is named."
          },
          "401": {
            "description": "No valid session."
          },
          "403": {
            "description": "Authenticated, but not permitted."
          }
        }
      }
    },
    "/api/v1/renewal-leads/{id}": {
      "get": {
        "tags": [
          "Renewal queue"
        ],
        "summary": "One renewal request, in full.",
        "description": "Needs `lead.read`. Carries the customer's contact details, because an operator picking it up has to be able to reach them; the policy number is masked.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed; the offending parameter is named."
          },
          "401": {
            "description": "No valid session."
          },
          "403": {
            "description": "Authenticated, but not permitted."
          }
        }
      },
      "patch": {
        "tags": [
          "Renewal queue"
        ],
        "summary": "Move a request along the workflow.",
        "description": "Needs `lead.write`. NEW → CONTACTED → QUOTE_REQUESTED / PARTNER_HANDOFF → CLOSED, forward only. NEW cannot jump to CLOSED: a request may only be closed once somebody has made contact. Closing requires a reason from a fixed set. Body is an allow-list — status, closedReason, assignToMe and nothing else.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed; the offending parameter is named."
          },
          "401": {
            "description": "No valid session."
          },
          "403": {
            "description": "Authenticated, but not permitted."
          }
        }
      }
    },
...
  }
}